+++ package.json
@@ -99,7 +99,6 @@
         "@typescript-eslint/eslint-plugin": "^5.4.0",
         "@typescript-eslint/parser": "^5.4.0",
         "babel-jest": "29.6.3",
-        "babel-plugin-module-resolver": "^5.0.0",
         "deepmerge": "3.2.0",
         "eslint": "8.19.0",
         "eslint-config-prettier": "8.8.0",
