+++ packages/adapter/README.md
@@ -1,0 +1,2 @@
+# @rnv/adapter
+
