+++ packages/adapter/tsconfig.json
@@ -1,0 +1,11 @@
+{
+    "extends": "@flexn/typescript-config/tsconfig.lib.node.json",
+    "compilerOptions": {
+        "outDir": "./lib",
+        "rootDir": "./src",
+        "resolveJsonModule": false,
+        "lib": ["ES2021.String"]
+    },
+    "include": ["./src/**/*", "./src/**/*.json", "./types.d.ts", "./package.json"],
+    "exclude": ["node_modules", "./src/**/*.test.ts"]
+}
