+++ packages/adapter/types.d.ts
@@ -1,0 +1,1 @@
+
