+++ packages/app-harness/metro.config.js
@@ -1,4 +1,4 @@
-const { withRNVMetro } = require('rnv');
+const { withRNVMetro } = require('@rnv/adapter');
 
 module.exports = withRNVMetro({
     resolver: {
