+++ packages/core/src/types.ts
@@ -13,32 +13,6 @@ export type RnvError = any;
 
 export type Env = Record<string, any>;
 
-export type BabelConfig = {
-    retainLines?: boolean;
-    plugins?: Array<
-        [
-            string,
-            {
-                root?: Array<string | undefined>;
-                alias?: Record<string, string | undefined>;
-            }
-        ]
-    >;
-    presets?: Array<
-        | string
-        | [
-              string,
-              {
-                  useTransformReactJSXExperimental?: boolean;
-              }
-          ]
-    >;
-};
-
-export type BabelApi = {
-    cache: (value: boolean) => void;
-};
-
 type Plat = Required<Required<ConfigFileBuildConfig>['platforms']>[PlatformKey];
 export type PlatPropKey = keyof Plat;
 export type BuildSchemePropKey = keyof Required<Plat>['buildSchemes'][string];
