+++ packages/engine-rn-macos/package.json
@@ -34,7 +34,8 @@
     "dependencies": {
         "@react-native/babel-preset": "0.73.21",
         "@rnv/sdk-apple": "1.0.0-rc.12",
-        "@rnv/sdk-react-native": "1.0.0-rc.12"
+        "@rnv/sdk-react-native": "1.0.0-rc.12",
+        "@rnv/adapter": "1.0.0-rc.12"
     },
     "peerDependencies": {
         "@rnv/core": "^1.0.0-rc.12"
