+++ packages/engine-rn-next/package.json
@@ -36,7 +36,8 @@
         "@rnv/sdk-utils": "1.0.0-rc.12",
         "babel-preset-expo": "9.5.2",
         "next-fonts": "1.5.1",
-        "next-images": "1.8.4"
+        "next-images": "1.8.4",
+        "@rnv/adapter": "1.0.0-rc.12"
     },
     "peerDependencies": {
         "@rnv/core": "^1.0.0-rc.12",
