+++ packages/engine-rn-windows/package.json
@@ -35,6 +35,7 @@
     },
     "dependencies": {
         "@rnv/sdk-react-native": "1.0.0-rc.12",
+        "@rnv/adapter": "1.0.0-rc.12",
         "@xmldom/xmldom": "0.7.7",
         "react-native-windows": "0.72.10"
     },
