+++ packages/engine-rn/package.json
@@ -35,7 +35,8 @@
         "@react-native/babel-preset": "0.73.21",
         "@rnv/sdk-android": "1.0.0-rc.12",
         "@rnv/sdk-apple": "1.0.0-rc.12",
-        "@rnv/sdk-react-native": "1.0.0-rc.12"
+        "@rnv/sdk-react-native": "1.0.0-rc.12",
+        "@rnv/adapter": "1.0.0-rc.12"
     },
     "peerDependencies": {
         "@rnv/core": "^1.0.0-rc.12"
