+++ packages/rnv/src/index.ts
@@ -1,2 +1,1 @@
-export * from './adapter';
 export * from './runner';
