+++ packages/template-starter/package.json
@@ -98,6 +98,7 @@
         "@flexn/graybox": "1.0.0-feat.12",
         "@lightningjs/sdk": "^5.4.1",
         "@rnv/core": "1.0.0-rc.12",
+        "@rnv/adapter": "1.0.0-rc.12",
         "@rnv/engine-lightning": "1.0.0-rc.12",
         "@rnv/engine-rn": "1.0.0-rc.12",
         "@rnv/engine-rn-electron": "1.0.0-rc.12",
