+++ packages/template-starter/react-native.config.js
@@ -1,5 +1,5 @@
 //! NO CONSOLE LOGS HERE. IT WILL BREAK THE APP
-const { withRNVRNConfig } = require('rnv');
+const { withRNVRNConfig } = require('@rnv/adapter');
 
 const config = withRNVRNConfig({});
