+++ packages/template-starter/renative.template.json
@@ -33,6 +33,7 @@
                 "@flexn/assets-renative-outline": "0.3.3",
                 "babel-loader": "9.1.3",
                 "@rnv/core": "1.0.0-rc.12",
+                "@rnv/adapter": "1.0.0-rc.12",
                 "rnv": "1.0.0-rc.12"
             },
             "browserslist": {
